Add files via upload
We were taught about logic gates in class but this task of designing it seemed really tough at the beginning. I actually didn't know how to work in TinkerCad and did have lot of confusions on how to do this task. But finally I could complete it. Hope its the way how to do it. I checked out a lot of references in youtube as well but gradually learnt how to design logic gates. Even now I am not fully aware of certain things that come as error in it but the fact is that I could learn something new and did my best to understand it.