Add files via upload
I am happy that I could do atleast one of the problem and could run it properly though i was not able to do the other problems. I tried a lot to solve the other problems in  task 3 but couldn't get the grip around the problems but I had the chance to get extra knowledge about each one of them. 
Some new things that i learnt include: 
int sum=0 which is done to avoid errors based on declaration of variables without initialization.
long long n
for loop usage - for iteration
etc;